Function wasm_chain_integration::constants::new_iterator_cost

pub fn new_iterator_cost(len: u32) -> u64

Cost of a new iterator. This accounts for tree traversal as well as the storage the execution engine needs to keep for the iterator. When looking up an iterator we construct a structure that keeps track of the current position in the tree. This iterator is constructed for any key, including the empty key. Hence the base cost of 80 is there to ensure we don’t run out of memory. It limits memory use to around 300MB in the worst case.

Additionally, since we have to store the key for the iterator we have to charge adequately so that memory use is bounded. This is the reason for the 100 factor.