Module wasm_chain_integration::utils

Various utilities for testing and extraction of schemas.

Structs

TestHost

A host which traps for any function call apart from report_error which it prints to standard out.

TrapHost

A host which traps for any function call.

Enums

ReportError

An auxiliary datatype used by report_error to be able to retain the structured information in case we want to use it later to insert proper links to the file, or other formatting.

WasmVersion

Functions

generate_contract_schema_v0

Tries to generate a state schema and schemas for parameters of methods of a V0 contract.

generate_contract_schema_v1

Tries to generate schemas for parameters and return values of methods for a contract with a V1 schema.

generate_contract_schema_v2

Tries to generate schemas for parameters and return values of methods for a contract with a V2 schema.

get_embedded_schema_v0

Get the embedded schema for smart contract modules version 0 if it exists.

get_embedded_schema_v1

Get the embedded schema for smart contract modules version 1 if it exists.

get_inits

Get the init methods of the module.

get_receives

Get the receive methods of the module.

run_module_tests

Instantiates the module with an external function to report back errors. Then tries to run exported test-functions, which are present if compile with the wasm-test feature.