Trait wasm_chain_integration::v1::trie::Collector

pub trait Collector<V> {
    fn add_value(&mut self, data: &V);

    fn add_path(&mut self, path: usize);

    fn add_children(&mut self, num_children: usize);
}

A type that can be used to collect auxiliary information while a mutable trie is being frozen. Particular use-cases of this are collecting the size of new data, as well as new persistent nodes.

Note that ideally these methods should return Result/Option so that we can terminate early in case of out of energy. However that is at the moment complex to set up with the way we exchange data through the FFI. Thus for now the behaviour is that we collect all the data, and then decide whether the collected size is too large. This is adequate since other energy bounds ensure that there is not too much data to collect, i.e., more than we could handle. Nevertheless it would be cleaner if we could terminate early.

Required methods

fn add_value(&mut self, data: &V)

Account for the stored value.

fn add_path(&mut self, path: usize)

Account for the stored path in the node (the “stem”). See low_level for an explanation of this concept. The argument is the length of the path in chunks, i.e., 4-bit segments.

fn add_children(&mut self, num_children: usize)

Account for storing the given number of children.